文档: 更换许可证为 Apache License 2.0
- 将项目许可证从 MIT License 更改为 Apache License 2.0
- Apache License 2.0 更适合企业级项目
- 提供更明确的专利保护条款
- 包含完整的许可证文本和版权信息

Copyright 2025 go-slim team

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.